Adjust numpy version constraint for windows compatibility
Numpy 2.0 requires pybind 2.12, but despite matching this requirement
the windows build seems to not work as the keras 3.3.3 library has
seemingly been compile with numpy 1.x support only.

Downgrading the version constraint to numpy < 2.0 fixes the issue for
now. To be noted, that issue does not exist with Linux (and maybe
macOS?) and this fixes should be revisited once official support for
Keras 3 is added.